pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "schema-scribe"
version = "0.2.0"
description = "AI-powered data documentation generator for databases and dbt projects."
readme = "README.md"
authors = [
    { name = "DongWon Moon", email = "ansehddnjs187@gmail.com" },
]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Documentation",
]

dependencies = [
    "typer>=0.9.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
    "ruamel.yaml>=0.17.0",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9.0"]
mariadb = ["mysql-connector-python>=8.0.0"]
google = ["google-generativeai>=0.4.0"]
ollama = ["ollama>=0.1.0"]
confluence = ["atlassian-python-api>=3.40.0"]
snowflake = ["snowflake-connector-python>=3.0.0"]
notion = ["notion-client>=2.0.0"]
server = ["fastapi>=0.100.0", "uvicorn[standard]>=0.20.0"]

test = [
    "pytest>=7.0.0",
    "pytest-mock>=3.0.0",
    "dbt-core>=1.7.0",
    "dbt-duckdb>=1.7.0",
]

all = [
    "psycopg2-binary>=2.9.0",
    "mysql-connector-python>=8.0.0",
    "google-generativeai>=0.4.0",
    "ollama>=0.1.0",
    "atlassian-python-api>=3.40.0",
    "snowflake-connector-python>=3.0.0",
    "notion-client>=2.0.0",
    "pytest>=7.0.0",
    "pytest-mock>=3.0.0",
    "dbt-core>=1.7.0",
    "dbt-duckdb>=1.7.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
]

[tool.setuptools.packages.find]
include = ["schema_scribe*"]
exclude = ["jaffle_shop_mini*", "tests*"]

[project.scripts]
schema-scribe = "schema_scribe.main:app"